Change title of the translation to “La forme en création”
I came up with this form of the title a couple of months ago,
but I mixed it up in my memory until I put the glossary in
writing and referred to my memory assistant notebook.

There is a subtlety in this title: beyond the choice of words,
detailed in RATIONALE sections in the glossary, using ‘en’ (in)
rather than ‘de la’ (of) creates a double entendre: the shape is
both the mold of creation as a discipline (shape applied to creation),
and the shape is a changing form in an ongoing creation process
(creation process applied to a shape).