Associate small-caps font variant with lowercase transform in headings
The lowercase transform used to obtain small capitals in TypeKit's font
ff-quadraat-sc-web lead to all-lowercase titles when the fallback font
Georgia is applied when the book is opened from localhost.

I added the small-caps style, which has no visible effect when the font
is available, and displays headings in small capitals in fallback font.