Start to rework long sentence
I changed my interpretation of this sentence, realizing that "Done well,"
at the very beginning is completed with "Our process will vary" at the
start of the next sentence.

I will join these two sentences by a comma in the French translation,
and move the block for "Done well" closer to avoid this long stretch
where the meaning gets lost.