Add cross-references with links between endnotes and copy
In the book, each endnote refers to a part of a sentence and the page
number where the sentence can be found. In the HTML version, no page
numbers are included. I added a link from each endnote to the corresponding
extract in the copy, as well as a link from the copy to the endnote.

To stay in the spirit of the book, I styled the anchors for these links
to be discreet, not visible until hover. Also, I highlighted the target
extract after a click, to make it easier to localize it in the page.