Remove TODO comment: not compatible with current styles
It would make sense to display the text of each chapter if the nav
was vertical rather than horizontal at narrow width in the responsive
view. With current layout, all the headings overlap.

The issue that I wished to address is that there is no hover effect
on mobile phones, since each tap of a finger results in a click. This
makes the nav rather cryptic. An alternative would be to display an
explicit table of contents in the index page, and move the foreword
to a separate page, foreword.html.