Add index.html, translated from English to French
I decided to translate the names of HTML pages as well;
while index.html remains the same, I updated the link
to opening.html within to refer to ouverture.html.

My rationale is that the file names appear in the URL
and are displayed in the status bar when hovering links,
and thus they should be meaningful to a French speaker.