Translate End Notes
I adapted the anchors and URLs to use French names:

  - chapter* -> chapitre*
  - endnotes.html -> notes.html

I also reformatted the notes slightly to make them look more familiar
to French readers.

I chose to include full English text of the original quotes in the notes,
and to leave English titles as is, followed with a French translation
when they are present in the main text and useful for the comprehension.