Replace little * at end of Chapter 10 with illustration of candles
The illustration of a candle lighting another candle appears on the
last page of the book. This is a reference to the haiku of Yosa Buson
quoted in Chapter 3, and the illustration appears in the bottom part
of the image which introduces Chapter 3.

This makes the end of the main copy more explicit.