Add French text of quote from Pascal
Modern French version taken from the Édition électronique
des Pensées de Blaise Pascal, by Dominique Descotes and
Gilles Proust.

http://www.penseesdepascal.fr/

This text differs from the English translation by one word,
'sphere' in English vs 'milieu' in French. The latter is found
in Pascal's manuscript. This somewhat changes the meaning:
we are sailing on a vast environment; the idea of the Earth
sphere floating in space is not present in the original.