Improve translation of the first sentence
I now translated "They say" by its French equivalent "On dit".
It is highly unusual to find this speech pattern in writing in French,
and emphasized like here at the beginning of a sentence, of a paragraph
and of a chapter, which is probably why I resisted putting it here in
the first place.

This is the most natural translation for this first sentence though,
half proverbial and half confidence. I emphasized the proverbial in
the second part of the sentence, which is now more concise.