Improve translation of 'loose gas'
I found that it is sometimes used as an equivalent of 'free gas',
as opposed to 'dissolved gas', the two forms of gas found in nature.

This hints at 'gaz libre' to translate 'loose gas'. This is probably
the correct expression from a technical point of view, and it is
equally valid in the pictorial evocation.