Improve translation of 'sealed nut' / 'open platter'
The previous translation did not feel natural in French.
I am now using a usual expression, 'assiette creuse', for 'open platter',
and I translated 'sealed' as 'pleine' (full) rather than 'fermée' (closed)
to keep a level opposition with 'creuse' (hollow).

The usage of the 'assiette creuse', or soup plate, is not limited to soup.
They are commonly used for service in informal settings, and even for
various dishes in restaurants, which makes it compatible with the
English word 'platter' in this context.