Translate paragraphs 23 and 24 of chapter 3
In the first item of the bullet list of paragraph 24, I translated
sizes from inches to centimeters, and wrote the dimensions as numbers
with abbreviated units rather than full text like the original because
the French translation is much longer than the English text, and I wished
to preserve the rhythm of the instructions, each on a single line.

Also, I rounded 3 inches (7.62cm) below, to 7 centimeters rather than 8,
to preserve the ratio of the two numbers: 20/3~7 and 50/7~7 but 50/8~6. And
7 is an odd number like 3, opposed to 20 which is an even number like 50.